Move `sierra_class_from_file` to `artifacts.rs`
